# Fix accidental O(n^2 * log n) performance in NixRepl::addAttrsToScope